chore: move isValidComponent & isValidMetaData
