+++ app/web-components/src/client/index.ts
@@ -1,6 +1,3 @@
-/* eslint-disable no-underscore-dangle */
-/* global window */
-
 export {
   storiesOf,
   setAddon,
@@ -12,21 +9,14 @@ export {
   raw,
 } from './preview';
 
-/**
- * @param customElements any for now as spec is not super stable yet
- */
-export function setCustomElements(customElements: any) {
-  // @ts-ignore
-  window.__STORYBOOK_CUSTOM_ELEMENTS__ = customElements;
-}
+export {
+  getCustomElements,
+  setCustomElements,
+  isValidComponent,
+  isValidMetaData,
+} from './customElements';
 
-export function getCustomElements() {
-  // @ts-ignore
-  return window.__STORYBOOK_CUSTOM_ELEMENTS__;
-}
-
+// TODO: disable HMR and do full page loads because of customElements.define
 if (module && module.hot && module.hot.decline) {
   module.hot.decline();
 }
-
-// TODO: disable HMR and do full page loads because of customElements.define
